+++ src/hardware/hardwareController.cpp
@@ -359,7 +359,7 @@ bool HardwareController::getVariableValue(string variable_name, float& value)
 
     if (variable_name == "Always")
     {
-        value = 1.0;
+        value = 1.f;
         return true;
     }
     if (variable_name == "HasShip")
@@ -368,16 +368,16 @@ bool HardwareController::getVariableValue(string variable_name, float& value)
         return true;
     }
     SHIP_VARIABLE("Hull", 100.0f * ship->hull_strength / ship->hull_max);
-    SHIP_VARIABLE("FrontShield", ship->getShieldPercentage(0));
-    SHIP_VARIABLE("RearShield", ship->getShieldPercentage(1));
-    SHIP_VARIABLE("Shield0", ship->getShieldPercentage(0));
-    SHIP_VARIABLE("Shield1", ship->getShieldPercentage(1));
-    SHIP_VARIABLE("Shield2", ship->getShieldPercentage(2));
-    SHIP_VARIABLE("Shield3", ship->getShieldPercentage(3));
-    SHIP_VARIABLE("Shield4", ship->getShieldPercentage(4));
-    SHIP_VARIABLE("Shield5", ship->getShieldPercentage(5));
-    SHIP_VARIABLE("Shield6", ship->getShieldPercentage(6));
-    SHIP_VARIABLE("Shield7", ship->getShieldPercentage(7));
+    SHIP_VARIABLE("FrontShield", static_cast<float>(ship->getShieldPercentage(0)));
+    SHIP_VARIABLE("RearShield", static_cast<float>(ship->getShieldPercentage(1)));
+    SHIP_VARIABLE("Shield0", static_cast<float>(ship->getShieldPercentage(0)));
+    SHIP_VARIABLE("Shield1", static_cast<float>(ship->getShieldPercentage(1)));
+    SHIP_VARIABLE("Shield2", static_cast<float>(ship->getShieldPercentage(2)));
+    SHIP_VARIABLE("Shield3", static_cast<float>(ship->getShieldPercentage(3)));
+    SHIP_VARIABLE("Shield4", static_cast<float>(ship->getShieldPercentage(4)));
+    SHIP_VARIABLE("Shield5", static_cast<float>(ship->getShieldPercentage(5)));
+    SHIP_VARIABLE("Shield6", static_cast<float>(ship->getShieldPercentage(6)));
+    SHIP_VARIABLE("Shield7", static_cast<float>(ship->getShieldPercentage(7)));
     SHIP_VARIABLE("Energy", ship->energy_level * 100 / ship->max_energy_level);
     SHIP_VARIABLE("ShieldsUp", ship->shields_active ? 1.0f : 0.0f);
     SHIP_VARIABLE("Impulse", ship->current_impulse * ship->getSystemEffectiveness(SYS_Impulse));
@@ -401,7 +401,7 @@ bool HardwareController::getVariableValue(string variable_name, float& value)
     for(int n=0; n<SYS_COUNT; n++)
     {
         SHIP_VARIABLE(getSystemName(ESystem(n)).replace(" ", "") + "Health", ship->systems[n].health);
-        SHIP_VARIABLE(getSystemName(ESystem(n)).replace(" ", "") + "Power", ship->systems[n].power_level / 3.0);
+        SHIP_VARIABLE(getSystemName(ESystem(n)).replace(" ", "") + "Power", ship->systems[n].power_level / 3.f);
         SHIP_VARIABLE(getSystemName(ESystem(n)).replace(" ", "") + "Heat", ship->systems[n].heat_level);
         SHIP_VARIABLE(getSystemName(ESystem(n)).replace(" ", "") + "Coolant", ship->systems[n].coolant_level);
         SHIP_VARIABLE(getSystemName(ESystem(n)).replace(" ", "") + "Hacked", ship->systems[n].hacked_level);
